+++ CHANGELOG.md
@@ -48,6 +48,7 @@
     - Fixed hanging pingpong-mode when changing slots manually
     - Added trigger-option "pseudo-random"
     - Added trigger-option "random walk"
+    - Added trigger-option "alternating"
 - Module [X4](./docs/X4.md)
     - Fixed advancing to the lower button after the upper button has been mapped
     - Fixed wrong tooltip of lower mapping button